Update support info generation
